app/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pymongo[srv]==4.6.0
python-dotenv==1.0.0
certifi==2023.11.17
openai==1.3.7
pydantic==2.5.0